fixed USB for s5l8930 with help from openiboot src
A couple of register differences in the include/plat files, some unknown
registers, PHY registers, and special values for FIFO size/address.

Tested on iPad1G with CONFIG_USB_ETH_RNDIS.

Some of the patching is a little dirty right now. It could be annoying
to merge with other kernels...