gpio-keys mostly works, on iPad at least
By mostly, I mean the volume buttons double-click for some reason. I
want to understand if there's something buggy about the GPIO interface
before just putting a debouncer on them.

There were a few issues:

* gpiolib was using IRQ number where it should use GPIO number instead
* the trigger types looked wrong and incomplete, compared to openiboot
and my tests with it
* demux code appeared to calculate the GPIO number incorrectly, though
I haven't tested with a high-enough GPIO to verify it's correct now
* the IRQ handler didn't correctly compute the IRQ number for VICs 2
and 3